ABI checker fixes part 3

+++ options/ansi/generic/string.cpp
@@ -288,7 +288,7 @@ int wcscmp(const wchar_t *l, const wchar_t *r) {
 
 int wcscoll(const wchar_t *, const wchar_t *) { MLIBC_STUB_BODY; }
 int wcsncmp(const wchar_t *, const wchar_t *, size_t) { MLIBC_STUB_BODY; }
-int wcsxfrm(wchar_t *__restrict, const wchar_t *__restrict, size_t) { MLIBC_STUB_BODY; }
+size_t wcsxfrm(wchar_t *__restrict, const wchar_t *__restrict, size_t) { MLIBC_STUB_BODY; }
 
 int wmemcmp(const wchar_t *a, const wchar_t *b, size_t size) {
 	for(size_t i = 0; i < size; i++) {
